Revert "Feature/code refactoring"
