--- Modules/MainModule/View/MovieListCell.swift
@@ -1,6 +1,0 @@
-import UIKit
-
-class MovieListCell: UITableViewCell {
-    @IBOutlet weak var movieTitle: UILabel!
-    @IBOutlet weak var movieDescription: UILabel!
-}
