--- Modules/MainModule/View/MovieResult.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-struct MovieResult{
-    let movieTitle: String
-    let movieDescription: String
-}
