+++ Modules/MainModule/View/ViewController.swift
@@ -1,48 +1,56 @@
 import UIKit
 
-class ViewController: UIViewController, UISearchBarDelegate, UITableViewDelegate, UITableViewDataSource{
+class ViewController: UIViewController, UISearchBarDelegate{
+    
     @IBOutlet weak var searchBar: UISearchBar!
-    
-    @IBOutlet weak var movieTable: UITableView!
-    
-    var listViewData: [MovieResult] = [
-        MovieResult(movieTitle: "Olá!", movieDescription: "Tente pesquisar por um filme.")
-    ]
+    var movieName: String?
+    var queryMovieName: String?
+    var queryString: String?
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
         searchBar.delegate = self
-        movieTable.dataSource = self
-        
-        
     }
     
     @IBAction func onButtonClick(_ sender: Any) {
-        MovieInteractor.fetchMovies(movieNameQuery: searchBar.text ?? ""){data in
-            guard let data = data else {
-                print("Failed to fetch data.")
-                return
-            }
-            self.listViewData = MoviePresenter.parseJSON(movieData: data)
-            
-            DispatchQueue.main.async{
-                self.movieTable.reloadData()
-            }
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return listViewData.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let item = listViewData[indexPath.row]
-        let cell = movieTable.dequeueReusableCell(withIdentifier: "MovieTableCell", for: indexPath) as! MovieListCell
+        movieName = searchBar.text ?? ""
+        queryMovieName = movieName?.replacingOccurrences(of: " ", with: "+")
+        queryString = "https://api.themoviedb.org/3/search/movie?query=\(queryMovieName!)&api_key=f33197e7d286b19bb9d55aeecca2975f"
         
-        cell.movieTitle.text = item.movieTitle
-        cell.movieDescription.text = item.movieDescription
-        return cell
+        if let url = URL(string: queryString!) {
+            let session = URLSession(configuration: .default)
+            
+            
+            let task = session.dataTask(with: url){data, urlResponse, error in
+                if error != nil{
+                    print(error!)
+                    return
+                } else {
+                    if let safeData = data {
+                        self.parseJSON(movieData: safeData)
+                    }
+                }
+            }
+            
+            task.resume()
+            
+        }
+        
+        
+    }
+    func parseJSON(movieData: Data){
+        let decoder = JSONDecoder()
+        do{
+            let decodedData = try decoder.decode(MovieData.self, from: movieData)
+            for movie in decodedData.results{
+                print(movie.title)
+                print(movie.overview)
+                print("---")
+            }
+        } catch {
+            print(error)
+        }
     }
     
 }
